Move Pin methods to atsamd51.
They are not chip-specific, they apply to the whole atsamd51 family.
